division matrix fix
